Implemented Search for Java Variable Names
Given a list of one million <string name, int score> pairs where names
are valid Java variable
names, write two programs and try to optimize their efficiency:
1. A Construction Program that produces index structure D.
2. A Query Server Program that reads in serialized D and then accepts
user queries such that for each query s, it responds with the top 10
names (ranked by score) that start with s or contains ‘_s’ (so for
example, both “revenue” and “yearly_revenue” match the prefix “rev”).
Query answering should run in sub-linear time (in terms of the number
of names in the input).